Enabled ShouldBeThrownBySpec.scala under Scala 3 build, enhanced matchers code to use Scala 3 macro for position to pass the tests.
