+++ jvm/matchers-core/src/main/scala/org/scalatest/matchers/dsl/ResultOfATypeInvocation.scala
@@ -20,6 +20,7 @@ import org.scalatest.Resources
 import org.scalatest.matchers.MatchersHelper.indicateSuccess
 import org.scalatest.matchers.MatchersHelper.indicateFailure
 import org.scalatest.matchers.MatchersHelper.checkThrownBy
+import org.scalatest.matchers.MatchersHelper.checkBeThrownBy
 import org.scalactic._
 
 import scala.reflect.ClassTag
@@ -83,23 +84,12 @@ final class ResultOfATypeInvocation[T](val clazzTag: ClassTag[T]) {
    *                              ^
    * </pre>
    **/
-  def should(beThrownBy: ResultOfBeThrownBy)(implicit prettifier: Prettifier, pos: source.Position): org.scalatest.Assertion = {
-    val throwables = beThrownBy.throwables
-    val noThrowable = throwables.find(_.isEmpty)
-    if (noThrowable.isDefined) {
-      val message = Resources.exceptionExpected(clazz.getName)
-      indicateFailure(message, None, pos)
-    }
-    else {
-      val unmatch = throwables.map(_.get).find(t => !clazz.isAssignableFrom(t.getClass))
-      if (unmatch.isDefined) {
-        val u = unmatch.get
-        val s = Resources.wrongException(clazz.getName, u.getClass.getName)
-        indicateFailure(s, Some(u), pos)
-      }
-      else indicateSuccess(Resources.exceptionThrown(clazz.getClass.getName))
-    }
-  }
+  // SKIP-DOTTY-START 
+  def should(beThrownBy: ResultOfBeThrownBy)(implicit prettifier: Prettifier, pos: source.Position): org.scalatest.Assertion = 
+    checkBeThrownBy(clazz, beThrownBy, pos)
+  // SKIP-DOTTY-END  
+  //DOTTY-ONLY inline def should(beThrownBy: ResultOfBeThrownBy)(implicit prettifier: Prettifier): org.scalatest.Assertion = 
+  //DOTTY-ONLY   ${ org.scalatest.matchers.MatchersHelper.checkBeThrownByMacro('{clazz}, '{beThrownBy}) }
 
   /**
    * This method enables the following syntax:
@@ -148,23 +138,12 @@ final class ResultOfATypeInvocation[T](val clazzTag: ClassTag[T]) {
    *                              ^
    * </pre>
    **/
-  def must(beThrownBy: ResultOfBeThrownBy)(implicit prettifier: Prettifier, pos: source.Position): org.scalatest.Assertion = {
-    val throwables = beThrownBy.throwables
-    val noThrowable = throwables.find(_.isEmpty)
-    if (noThrowable.isDefined) {
-      val message = Resources.exceptionExpected(clazz.getName)
-      indicateFailure(message, None, pos)
-    }
-    else {
-      val unmatch = throwables.map(_.get).find(t => !clazz.isAssignableFrom(t.getClass))
-      if (unmatch.isDefined) {
-        val u = unmatch.get
-        val s = Resources.wrongException(clazz.getName, u.getClass.getName)
-        indicateFailure(s, Some(u), pos)
-      }
-      else indicateSuccess(Resources.exceptionThrown(clazz.getClass.getName))
-    }
-  }
+  // SKIP-DOTTY-START 
+  def must(beThrownBy: ResultOfBeThrownBy)(implicit prettifier: Prettifier, pos: source.Position): org.scalatest.Assertion = 
+    checkBeThrownBy(clazz, beThrownBy, pos)
+  // SKIP-DOTTY-END
+  //DOTTY-ONLY inline def must(beThrownBy: ResultOfBeThrownBy)(implicit prettifier: Prettifier): org.scalatest.Assertion = 
+  //DOTTY-ONLY   ${ org.scalatest.matchers.MatchersHelper.checkBeThrownByMacro('{clazz}, '{beThrownBy}) }
   
   override def toString: String = "a [" + clazz.getName + "]"
 }
